# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amar Steps - Swagger DOC ",
        "description": "L6 Swagger OpenApi description",
        "contact": {
            "email": "nshakib.se@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "127.0.0.1:8000",
            "description": "Sandbox OpenApi dynamic host server"
        }
    ],
    "paths": {
        "/api/v1/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Get auth activities",
                "description": "Returns auth data",
                "operationId": "activities-get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve auth activities.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Save activity log.",
                "description": "Activity has been created successfully.",
                "operationId": "activities-post",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "query",
                        "description": "activity json",
                        "required": true,
                        "schema": {
                            "type": "json"
                        }
                    },
                    {
                        "name": "notes",
                        "in": "query",
                        "description": "notes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity has been created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": true,
                                        "data": [
                                            "saved data"
                                        ],
                                        "message": "Activity has been created successfully."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Prerequisite failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": false,
                                        "data": [
                                            "errors"
                                        ],
                                        "message": "Prerequisite failed."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    }
                }
            }
        },
        "/api/v1/activities/{id}": {
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Delete activity.",
                "description": "Activity has been deleted successfully.",
                "operationId": "destroy-activity",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity has been deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": true,
                                        "data": "",
                                        "message": "Activity has been deleted successfully."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Failed to delete.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": false,
                                        "message": "Failed to delete."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/badges": {
            "get": {
                "tags": [
                    "Badges"
                ],
                "summary": "Get all badges",
                "description": "Returns badges",
                "operationId": "badges-get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve badges.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Badges"
                ],
                "summary": "Create badge.",
                "description": "Badge has been created successfully.",
                "operationId": "Badges-post",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "display_name",
                        "in": "query",
                        "description": "display_name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Badge has been created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": true,
                                        "data": [
                                            "saved data"
                                        ],
                                        "message": "Badge has been created successfully."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Prerequisite failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": false,
                                        "data": [
                                            "errors"
                                        ],
                                        "message": "Prerequisite failed."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}